verify-mfa/requirements.txt
pyotp
mysql-connector-python
cryptography
requests
qrcode